создал 6 тестов к методу distince

+++ lesson_1/src/main/java/qa/paket2/Point.java
@@ -17,7 +17,7 @@ public class Point {
   public double distance(Point p1,Point p2){
     double rezx = p1.x-p2.x;
     double rezy = p1.y-p2.y;
-    double z = Math.sqrt(rezx*rezx+rezy*rezy);
+    int z = (int) Math.sqrt(rezx*rezx+rezy*rezy);
     System.out.println("The distance between 2 points equals = " + z);
     return z;
   }
